FIX: Unittests
Alongside: Many methods that used incorrect foreign keys.

+++ amsterdam_app_api/GarbageCollector/GarbageCollector.py
@@ -16,37 +16,32 @@ class GarbageCollector:
     def collect_iprox(self, project_type=None):
         """ Call all iprox dataset garbage collectors one by one """
         report_projects = {}
+        report_project_details = {}
+        report_news = {}
 
         # If there's no path, projects might be de-activated un-rightfully
         if project_type is not None:
+            news = list(News.objects.filter(project_type=project_type).all())
+            report_news = self.garbage_collect_iprox(news, model=News, model_name='News')
+
+            project_details = list(ProjectDetails.objects.filter(project_type=project_type).all())
+            report_project_details = self.garbage_collect_iprox(project_details, model=ProjectDetails)
+
             projects = list(Projects.objects.filter(project_type=project_type).all())
-            report_projects = self.garbage_collect_iprox(projects, model=Projects, siblings=True)
+            report_projects = self.garbage_collect_iprox(projects, model=Projects, model_name='Projects')
 
-        return {'projects': report_projects}
+        return {'projects': report_projects, 'project_details': report_project_details, 'news': report_news}
 
-    def garbage_collect_iprox(self, data_objects, model=None, model_name=None, siblings=False):
+    def garbage_collect_iprox(self, data_objects, model=None, model_name=None):
         """ Generic garbage collections method """
         report = {}
         for data_object in data_objects:
-            # DEBUG
             # Check if we've seen this project anywhere in the last week, if not -> delete!
             if data_object.last_seen + timedelta(days=7) <= self.last_scrape_time:
-                # if siblings is True:
-                #     # Remove ProjectDetails, News, Notifications, Warning-messages
-                #     ProjectDetails.objects.filter(identifier=data_object.identifier).delete()
-                #     News.objects.filter(project_identifier=data_object.identifier).delete()
-                #     Notification.objects.filter(project_identifier=data_object.identifier).delete()
-                #     WarningMessages.objects.filter(project_identifier=data_object.identifier).delete()
-                #     project_managers = list(ProjectManager.objects.all())
-                #
-                #     # Cleanup authorizations
-                #     for project_manager in project_managers:
-                #         if data_object.identifier in project_manager.projects:
-                #             projects = project_manager.projects.remove(data_object.identifier)
-                #             data = {'projects': projects if projects is not None else []}
-                #             ProjectManager.objects.filter(pk=project_manager.identifier).update(**data)
                 report[data_object.identifier] = 'deleted'
-                data_object.delete()
+                if model_name == 'Projects':
+                    # Siblings are cascaded deleted
+                    data_object.delete()
 
             # We've seen the object (again!), mark active=true
             elif data_object.last_seen >= self.last_scrape_time:
